Move Unit's stringToUnitMap out of logic files
Deciding what specific strings should turn into which units is something
for the user code to decide, not the logic part of the code. The map was
moved to a class which is instantiated in the main function instead of
instantiating it directly in main to avoid cluttering the code with a
huge initialization statement. It also prevents the
constructor/destructor from executing at program startup/shutdown.