Move specialized convert_ functions out of impl
The convert_* functions now accept separate enum types as arguments, and
as such can be overloaded as simply convert. If the user has two
enumerations of the same type it makes sense to just allow them to call
the specialized convert function directly instead of having to create
Units just for convert to confirm that they have the same enum type,
which we already knew from the start.