Terminate the program if an unreachable branch is taken
Even if switches over an enum class cover every enumerator, someone
could technically pass an invalid enumeration as the argument.
Additionally, if a new enumerator is added, it's possible to forget to
add a new case for the switch. In either such case, the program would
invoke undefined behavior by not returning anything before exiting the
function. To avoid this, I just add a call to std::terminate in the
branches that should be unreachable assuming the arguments passed to the
functions are valid and that every enumerator is covered in the
switches.