Add a convert overload that uses strings
The plan is to let the GUI version work in the same way as the shell version if
arguments are provided to it. In order to avoid duplicating code in both
jconverter-gui.cpp and jconverter-shell.cpp, I want them to share a function
that converts units based on the arguments provided.